+++ src/ADC.h
@@ -1,7 +1,9 @@
 #ifndef __ADC_H__
 #define __ADC_H__
 
+#include <stdint.h>
+
 void adc_init(void);
-unsigned int get_adc(unsigned int p);
+uint16_t get_adc(uint16_t p);
 
 #endif
